Several changes! Combat implemented, enemies lootable. So far only two enemies in room 2 for testing purposes. Currently, several debug lines still scattered around the code.
Multiple methods were migrated to the controller 'Adventure'.

Moreover, I was having issues with completing my UI-Loop. Ultimately, I reversed the logic of events and included a combat and general options enums as well.

We now start the loop with a conditional check for 'inCombat'. If '!isInCombat' we skip to relevant logic for exploration.